moving pagination logic to shared util
